Correction bugs

+++ app/src/main/java/com/g/laurent/backtobike/Controllers/Activities/MainActivity.java
@@ -2,7 +2,6 @@ package com.g.laurent.backtobike.Controllers.Activities;
 
 import android.app.Dialog;
 import android.app.FragmentTransaction;
-import android.content.Context;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
@@ -13,7 +12,6 @@ import android.widget.EditText;
 import android.widget.ProgressBar;
 import android.widget.Toast;
 import com.g.laurent.backtobike.Controllers.Fragments.MainFragment;
-import com.g.laurent.backtobike.Models.AppDatabase;
 import com.g.laurent.backtobike.Models.CallbackCounters;
 import com.g.laurent.backtobike.Models.CallbackMainActivity;
 import com.g.laurent.backtobike.Models.CallbackSynchronizeEnd;
@@ -31,7 +29,6 @@ import com.google.android.gms.maps.model.LatLng;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import java.util.List;
-import static com.g.laurent.backtobike.Utils.UtilsApp.areCharactersAllowed;
 
 
 public class MainActivity extends BaseActivity implements CallbackMainActivity {
@@ -46,6 +43,11 @@ public class MainActivity extends BaseActivity implements CallbackMainActivity {
     private static final String BUNDLE_DIFFERENCES = "bundle_differences";
     private FirebaseUser user;
     private ProgressBar progressBar;
+    private List<Difference> differenceList;
+    private  List<String> listDifferencesFriendsAndInvits;
+    private int counterFriend;
+    private int counterEvents;
+    private int counterInvits;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -68,7 +70,7 @@ public class MainActivity extends BaseActivity implements CallbackMainActivity {
         UtilsTime.deleteOverdueEvents(getApplicationContext(), userId);
 
         // Check if the database has already been initialized (during the first use of the app on the phone)
-        if(userId!=null)
+        if(userId!=null && UtilsApp.isInternetAvailable(getApplicationContext()))
             checkInitializationDatabase();
         else
             getCurrentLocationForWeather();
@@ -77,7 +79,7 @@ public class MainActivity extends BaseActivity implements CallbackMainActivity {
     @Override
     public void onResume() {
         super.onResume();
-        if(userId!=null) {
+        if(userId!=null && UtilsApp.isInternetAvailable(getApplicationContext())) {
             UtilsTime.deleteOverdueEvents(getApplicationContext(), userId);
             defineCountersAndConfigureToolbar(MENU_MAIN_PAGE);
         }
@@ -88,10 +90,6 @@ public class MainActivity extends BaseActivity implements CallbackMainActivity {
         defineCountersAndConfigureToolbar(MENU_MAIN_PAGE);
     }
 
-    // --------------------------------------------------------------------------------------------------------
-    // -------------------------------------- LOGIN CHECKING --------------------------------------------------
-    // --------------------------------------------------------------------------------------------------------
-
     private void checkInitializationDatabase(){
 
         if(!sharedPref.getBoolean(SHAREDPREFERENCES_INIT,false)){
@@ -100,21 +98,66 @@ public class MainActivity extends BaseActivity implements CallbackMainActivity {
                 SynchronizeWithFirebase.buildDatabaseWithDatasFromFirebase(userId, sharedPref, getApplicationContext(), new CallbackSynchronizeEnd() {
                     @Override
                     public void onCompleted() {
-                        checkIfUserHasLoginOnFirebase();
+                        defineCountersAndConfigureToolbar(MENU_MAIN_PAGE);
                     }
 
                     @Override
                     public void onFailure(String error) {
-                        checkIfUserHasLoginOnFirebase();
+                        defineCountersAndConfigureToolbar(MENU_MAIN_PAGE);
                     }
                 });
             } finally {
-                checkIfUserHasLoginOnFirebase();
+                defineCountersAndConfigureToolbar(MENU_MAIN_PAGE);
             }
         } else {
-            checkIfUserHasLoginOnFirebase();
+            defineCountersAndConfigureToolbar(MENU_MAIN_PAGE);
         }
     }
+
+    // --------------------------------------------------------------------------------------------------------
+    // ---------------------------------- 1 - GET COUNTERS INFO -----------------------------------------------
+    // --------------------------------------------------------------------------------------------------------
+
+    public void defineCountersAndConfigureToolbar(String typeDisplay){
+
+        if(userId!=null && UtilsApp.isInternetAvailable(getApplicationContext())){
+            FirebaseRecover firebaseRecover = new FirebaseRecover(getApplicationContext());
+            firebaseRecover.recoverDatasForCounters(userId, getApplicationContext(), new CallbackCounters() {
+                @Override
+                public void onCompleted(List<Difference> differenceList, List<String> listDifferencesFriendsAndInvits, int counterFriend, int counterEvents, int counterInvits) {
+                    toolbarManager.configureToolbar(callbackBaseActivity, typeDisplay, counterFriend, counterEvents, counterInvits);
+                    int count = counterFriend + counterEvents + counterInvits;
+                    UtilsApp.setBadge(getApplicationContext(), count);
+                    setCounters(differenceList, listDifferencesFriendsAndInvits, counterFriend, counterEvents, counterInvits);
+                }
+
+                @Override
+                public void onFailure(String error) {
+                    toolbarManager.configureToolbar(callbackBaseActivity, typeDisplay, 0, 0,0);
+                    UtilsApp.setBadge(getApplicationContext(), 0);
+                    setCounters(null, null, 0,0,0);
+                }
+            });
+        } else {
+            setCounters(null, null, 0,0,0);
+        }
+
+    }
+
+    private void setCounters(List<Difference> differenceList, List<String> listDifferencesFriendsAndInvits, int counterFriend, int counterEvents, int counterInvits){
+
+        this.differenceList=differenceList;
+        this.listDifferencesFriendsAndInvits = listDifferencesFriendsAndInvits;
+        this.counterFriend=counterFriend;
+        this.counterEvents=counterEvents;
+        this.counterInvits=counterInvits;
+
+        checkIfUserHasLoginOnFirebase();
+    }
+
+    // --------------------------------------------------------------------------------------------------------
+    // ------------------------------------- 2 - LOGIN CHECKING -----------------------------------------------
+    // --------------------------------------------------------------------------------------------------------
 
     private void checkIfUserHasLoginOnFirebase(){
 
@@ -127,8 +170,19 @@ public class MainActivity extends BaseActivity implements CallbackMainActivity {
                     // Save login in sharedpreferences
                     sharedPref.edit().putString(LOGIN_SHARED,login).apply();
 
-                    // Get current location
-                    getCurrentLocationForWeather();
+                    // Synchronize events
+                    SynchronizeWithFirebase.synchronizeMyEvents(userId, getApplicationContext(), new CallbackSynchronizeEnd() {
+                        @Override
+                        public void onCompleted() {
+                            // Get current location
+                            getCurrentLocationForWeather();
+                        }
+
+                        @Override
+                        public void onFailure(String error) {
+                            Toast.makeText(getApplicationContext(),error,Toast.LENGTH_LONG).show();
+                        }
+                    });
 
                 } else { // login needs to be provided by user
                     showDialogToDefineLogin(); // show dialog
@@ -155,10 +209,7 @@ public class MainActivity extends BaseActivity implements CallbackMainActivity {
         Button dialogButtonSave = dialog.findViewById(R.id.button_save);
         dialogButtonSave.setOnClickListener(v -> {
             if(login.getText().length()>0) {
-                if(areCharactersAllowed(login.getText().toString()))
-                    checkLogin(login.getText().toString(), dialog);
-                else
-                    Toast.makeText(getApplicationContext(), getApplicationContext().getResources().getString(R.string.error_forbidden_characters), Toast.LENGTH_LONG).show();
+                checkLogin(login.getText().toString(), dialog);
             } else
                 Toast.makeText(getApplicationContext(),getApplicationContext().getResources().getString(R.string.error_login),Toast.LENGTH_LONG).show();
         });
@@ -185,13 +236,25 @@ public class MainActivity extends BaseActivity implements CallbackMainActivity {
                 } else {
                     FirebaseUpdate firebaseUpdate = new FirebaseUpdate(getApplicationContext());
                     firebaseUpdate.updateUserData(userId, user.getDisplayName(), user.getPhotoUrl().toString(),login);
+
                     dialog.dismiss();
 
                     // Save login in sharedpreferences
                     sharedPref.edit().putString(LOGIN_SHARED,login).apply();
 
-                    // Get current location
-                    getCurrentLocationForWeather();
+                    // Synchronize events
+                    SynchronizeWithFirebase.synchronizeMyEvents(userId, getApplicationContext(), new CallbackSynchronizeEnd() {
+                        @Override
+                        public void onCompleted() {
+                            // Get current location
+                            getCurrentLocationForWeather();
+                        }
+
+                        @Override
+                        public void onFailure(String error) {
+                            Toast.makeText(getApplicationContext(),error,Toast.LENGTH_LONG).show();
+                        }
+                    });
                 }
             }
 
@@ -203,12 +266,14 @@ public class MainActivity extends BaseActivity implements CallbackMainActivity {
     }
 
     // --------------------------------------------------------------------------------------------------------
-    // -------------------------------------- WEATHER METHODS -------------------------------------------------
+    // ------------------------------------ 3 - WEATHER METHODS -----------------------------------------------
     // --------------------------------------------------------------------------------------------------------
 
     public void getCurrentLocationForWeather(){
         GetCurrentLocation getCurrentLocation = new GetCurrentLocation();
-        getCurrentLocation.getLocationPermission(this, sharedPref, currentLocation -> defineCountersAndConfigureToolbar(currentLocation, MENU_MAIN_PAGE));
+        getCurrentLocation.getLocationPermission(this, sharedPref, currentLocation ->
+                configureMainFragment(UtilsCounters.transformListDifferencesToString(differenceList,
+                        listDifferencesFriendsAndInvits),counterEvents, counterFriend, counterInvits, currentLocation));
     }
 
     @Override
@@ -230,36 +295,8 @@ public class MainActivity extends BaseActivity implements CallbackMainActivity {
     }
 
     // --------------------------------------------------------------------------------------------------------
-    // -------------------------------------- CONFIGURE VIEWS -------------------------------------------------
+    // ----------------------------------- 4 - CONFIGURE VIEWS ------------------------------------------------
     // --------------------------------------------------------------------------------------------------------
-
-    public void defineCountersAndConfigureToolbar(LatLng currentLocation, String typeDisplay){
-
-        if(userId!=null && UtilsApp.isInternetAvailable(getApplicationContext())){
-            FirebaseRecover firebaseRecover = new FirebaseRecover(getApplicationContext());
-            firebaseRecover.recoverDatasForCounters(userId, getApplicationContext(), new CallbackCounters() {
-                @Override
-                public void onCompleted(List<Difference> differenceList, List<String> listDifferencesFriendsAndInvits, int counterFriend, int counterEvents, int counterInvits) {
-                    toolbarManager.configureToolbar(callbackBaseActivity, typeDisplay, counterFriend, counterEvents, counterInvits);
-                    int count = counterFriend + counterEvents + counterInvits;
-                    UtilsApp.setBadge(getApplicationContext(), count);
-
-                    configureMainFragment(UtilsCounters.transformListDifferencesToString(differenceList, listDifferencesFriendsAndInvits),
-                            counterEvents, counterFriend, counterInvits, currentLocation);
-                }
-
-                @Override
-                public void onFailure(String error) {
-                    toolbarManager.configureToolbar(callbackBaseActivity, typeDisplay, 0, 0,0);
-                    UtilsApp.setBadge(getApplicationContext(), 0);
-                    configureMainFragment(null,0,0,0, currentLocation);
-                }
-            });
-        } else {
-            configureMainFragment(null,0,0,0, currentLocation);
-        }
-
-    }
 
     private void configureMainFragment(String differences, int countEvent, int countFriends, int countInvits, LatLng currentLocation){
 
@@ -293,16 +330,5 @@ public class MainActivity extends BaseActivity implements CallbackMainActivity {
 
     public SharedPreferences getSharedPref() {
         return sharedPref;
-    }
-
-
-    private void clearDatabase(String userId, Context context){
-
-        AppDatabase.getInstance(context,userId).eventFriendsDao().deleteAllEventFriends();
-        //AppDatabase.getInstance(context,userId).friendsDao().deleteAllFriends();
-        AppDatabase.getInstance(context,userId).routeSegmentDao().deleteRouteSegment();
-        AppDatabase.getInstance(context,userId).routesDao().deleteAllRoutes();
-        AppDatabase.getInstance(context,userId).bikeEventDao().deleteAllBikeEvents();
-
     }
 }
